Real implementation of the ios plugin
The ios plugin was a bit more complicated than the android plugin. Loading the
new view controller was easier, of course - we can just look it up by nib name.
But simply showing the new view controller was not sufficient - since the
parent of the cordova controller is not a nav controller, showing the new view
replaced the existing view, which meant that there was no way to navigate back.

I tried all the techniques for displaying a new view (showViewController,
showDetailViewController, presentViewController), with a bunch of modal
presentation styles, and none of them provided a graceful way to transition
back.

So I worked around this by adding a navigationController as a parent of the
current view controller, and pushing the settings view controller onto it.

This means that the native navbar hides the cordova navbar, so when we navigate
back to the main screen, I hide the native navbar.

An alternative approach would be to require that the native screen have a
mechanism to dismiss itself. Since I wanted to ensure that this works with
arbitrary native screens, I chose this approach, at the cost of making the
cordova view controller a child of a nav controller.

I am not sure why the view controller is not the child of a nav controller to
begin with since that would make true hybrid apps so much easier.